Start of project
Title: Exploratory data analysis and Visualization of face masks data (Deadline : 20th July 2024)

Description:

For this assignment, your objective will be to analyze two datasets related to face masks and provide insights to a manufacturer of personal care products. The manufacturer wants to understand the competition and improve their marketing strategy and the research and development department. The datasets include a list of facemasks available on an online store along with information about the product, price, ratings, and list of consumer reviews for these face masks.

Questions:

What are the most popular face masks in the store and what are their key features (e.g. price, material, design)?

What do consumers like about the most popular face masks and why?

What are the different profiles of consumers who buy face masks and how can we segment them based on their demographics, behaviors, and preferences?

How can additional data sources be used to gain more insights into the type of consumers buying different face masks?

What are the key findings and implications of this analysis for the client's marketing strategy and research and development department?

What are the potential constraints and improvements to the analysis and how can we address them?

Identify and suggest a few extra features which could have helped you to gain a few more valuable insights into the different types of consumers who purchase these face masks.

Deliverables:

A visual presentation of the analysis approach, methodology, and key findings